@@ -3,93 +3,85 @@
     "================================================================================",
     "# To work within the development environment, run the following tasks",
     "  1. $ npm run dev",
-    "  2. $ npm run serve",
-    "  3. Go to http://localhost:8000",
+    "  2. Go to http://localhost:8000",
     "================================================================================"
   ],
   "name": "@aurodesignsystem/auro-lockup",
-  "version": "2.0.2",
+  "version": "0.0.0",
   "description": "auro-lockup HTML custom element",
   "repository": {
     "type": "git",
     "url": "https://github.com/AlaskaAirlines/auro-lockup"
   },
+  "type": "module",
   "main": "index.js",
   "license": "Apache-2.0",
+  "engines": {
+    "node": "^18.x || ^20.x "
+  },
   "dependencies": {
-    "@alaskaairux/icons": "^4.6.0",
-    "chalk": "^4.1.1",
-    "lit-element": "^2.5.1"
+    "chalk": "^5.3.0",
+    "lit": "^3.1.1"
   },
   "peerDependencies": {
-    "@alaskaairux/design-tokens": "^3.3.0",
-    "@alaskaairux/webcorestylesheets": "^3.4.0",
-    "@webcomponents/webcomponentsjs": "^2.5.0",
-    "focus-visible": "^5.2.0"
+    "@aurodesignsystem/design-tokens": "^4.3.0",
+    "@aurodesignsystem/webcorestylesheets": "^5.0.5"
   },
   "devDependencies": {
-    "@alaskaairux/design-tokens": "^3.3.0",
-    "@alaskaairux/webcorestylesheets": "^3.4.0",
-    "@aurodesignsystem/eslint-config": "^1.1.0",
-    "@babel/core": "^7.14.6",
-    "@babel/plugin-syntax-dynamic-import": "^7.8.3",
-    "@babel/plugin-transform-runtime": "^7.14.5",
-    "@babel/preset-env": "^7.14.5",
-    "@babel/runtime": "^7.14.6",
-    "@commitlint/cli": "^12.1.4",
-    "@commitlint/config-conventional": "^12.1.4",
-    "@open-wc/testing": "^2.5.33",
-    "@rollup/plugin-alias": "^3.1.2",
-    "@rollup/plugin-babel": "^5.3.0",
-    "@rollup/plugin-commonjs": "^19.0.0",
-    "@rollup/plugin-node-resolve": "^13.0.0",
-    "@semantic-release/changelog": "^5.0.1",
-    "@semantic-release/git": "^9.0.0",
-    "@semantic-release/npm": "^7.1.3",
-    "@web/dev-server": "^0.1.17",
-    "@web/test-runner": "^0.13.11",
-    "@webcomponents/webcomponentsjs": "^2.5.0",
-    "autoprefixer": "^10.2.6",
-    "babel-loader": "^8.2.2",
+    "@alaskaairux/icons": "^4.32.0",
+    "@aurodesignsystem/auro-library": "^2.2.6",
+    "@aurodesignsystem/design-tokens": "^4.3.0",
+    "@aurodesignsystem/eslint-config": "^1.3.0",
+    "@aurodesignsystem/webcorestylesheets": "^5.0.5",
+    "@commitlint/cli": "^18.4.4",
+    "@commitlint/config-conventional": "^18.4.4",
+    "@open-wc/testing": "^4.0.0",
+    "@rollup/plugin-node-resolve": "^15.2.3",
+    "@semantic-release/changelog": "^6.0.3",
+    "@semantic-release/git": "^10.0.1",
+    "@semantic-release/npm": "^11.0.2",
+    "@web/dev-server": "^0.4.1",
+    "@web/test-runner": "^0.18.0",
+    "autoprefixer": "^10.4.17",
     "compression": "^1.7.4",
     "concat": "^1.0.3",
+    "concurrently": "^8.2.2",
     "copyfiles": "^2.4.1",
-    "core-js": "^3.14.0",
-    "eslint": "^7.28.0",
-    "eslint-plugin-lit": "^1.5.1",
-    "focus-visible": "^5.2.0",
-    "husky": "^6.0.0",
+    "core-js": "^3.35.0",
+    "eslint": "^8.56.0",
+    "eslint-plugin-jsdoc": "^48.0.2",
+    "husky": "^8.0.0",
     "lodash": "^4.17.21",
-    "marked": "^2.1.0",
-    "nodemon": "^2.0.7",
+    "markdown-magic": "^2.6.1",
+    "nodemon": "^3.0.3",
     "npm-run-all": "^4.1.5",
-    "postcss": "^8.3.4",
-    "postcss-custom-properties": "^11.0.0",
-    "postcss-discard-comments": "^5.0.1",
-    "postcss-rem-to-pixel": "^4.1.2",
-    "postcss-remove-rules": "^1.0.0",
+    "postcss": "^8.4.33",
+    "postcss-discard-comments": "^6.0.1",
     "postcss-selector-replace": "^1.0.2",
-    "prismjs": "^1.23.0",
-    "rollup": "^2.52.0",
-    "rollup-plugin-minify-html-literals": "^1.2.6",
-    "rollup-plugin-serve": "^1.1.0",
-    "rollup-plugin-terser": "^7.0.2",
-    "semantic-release": "^17.4.4",
-    "sass": "^1.35.1",
-    "stylelint": "^13.13.1",
-    "stylelint-config-idiomatic-order": "^8.1.0",
-    "stylelint-config-standard": "^22.0.0",
-    "stylelint-order": "^4.1.0",
-    "stylelint-scss": "^3.19.0",
-    "wc-sass-render": "^1.2.3",
-    "web-component-analyzer": "^1.1.6",
-    "whatwg-fetch": "^3.6.2",
-    "yaml-lint": "^1.2.4"
+    "rollup": "^4.9.5",
+    "rollup-plugin-serve": "^1.1.1",
+    "sass": "^1.70.0",
+    "semantic-release": "^23.0.0",
+    "stylelint": "^16.2.0",
+    "stylelint-config-idiomatic-order": "^10.0.0",
+    "stylelint-config-standard": "^36.0.0",
+    "stylelint-config-standard-scss": "^13.0.0",
+    "stylelint-order": "^6.0.4",
+    "stylelint-scss": "^6.0.0",
+    "typescript": "^5.3.3",
+    "wc-sass-render": "^1.4.0",
+    "web-component-analyzer": "^2.0.0",
+    "whatwg-fetch": "^3.6.20",
+    "yaml-lint": "^1.7.0"
+  },
+  "overrides": {
+    "wc-sass-render": {
+      "sass": "$sass"
+    }
   },
   "browserslist": [
     "last 2 Chrome versions",
     "last 2 iOS major versions",
-    "ie 11",
     "last 2 Firefox versions",
     "last 2 Edge versions",
     "last 2 Safari major versions"
